Fixed case on default mode and disabled mode
VSCode is giving a squigly yellow line on the "Default" value of the "C_Cpp.intelliSenseEngine": "Default" key on the *.code-workspace file. Not sure if this is the proper fix or even if this is a mac only thing.
The error says:
Value is not accepted. Valid values: "default", "Tag Parser", "disabled".